Tp2_MongoDB/requirements.txt
pymongo>=4.3
python-dotenv>=1.0
dnspython>=2.3
requests>=2.28
psycopg2-binary>=2.9


